core/backend/requirements.txt
ffmpeg-python>=0.2.0
requests>=2.31
Flask>=3.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
debugpy>=1.8
